backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
python-multipart==0.0.9
SQLAlchemy==2.0.35
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg[binary]==3.1.19
alembic==1.13.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pyjwt[crypto]==2.9.0
itsdangerous==2.2.0
ldap3==2.9.1
